Added Classification Feature for Deliverable 2
Completed adding classification feature to .ics program. User is
prompted and asked whether or not they want to enter a classification,
if they do it will allow them to input their choice. If not it will
default to PUBLIC.

Basic error checking was added but nothing crazy, should work as
explained.