Deliverable 3 reading in an ICS file
Under the main method I started to add methods that can be used for
deliverable 3.

This one is the method that will allow a user to input any number of
ICS files and have them put into an arrayList. Should be able to pretty
much copy and paste the code into any of our stuff in order to use it.